+++ sw/airborne/stm32/can_hw.h
@@ -26,6 +26,6 @@
 #define CAN_HW_H
 
 void can_hw_init(void);
-int can_hw_transmit(uint16_t id, const uint8_t *buf, uint8_t len);
+int can_hw_transmit(uint32_t id, const uint8_t *buf, uint8_t len);
 
 #endif /* CAN_HW_H */
